+++ src/sentry_core.c
@@ -502,6 +502,22 @@ static
     return send;
 }
 
+static const char *
+str_from_attachment_type(sentry_attachment_type_t attachment_type)
+{
+    switch (attachment_type) {
+    case ATTACHMENT:
+        return "event.attachment";
+    case MINIDUMP:
+        return "event.minidump";
+    case VIEW_HIERARCHY:
+        return "event.view_hierarchy";
+    default:
+        UNREACHABLE("Unknown attachment type");
+        return "event.attachment";
+    }
+}
+
 sentry_envelope_t *
 sentry__prepare_event(const sentry_options_t *options, sentry_value_t event,
     sentry_uuid_t *event_id, bool invoke_before_send)
@@ -544,6 +560,15 @@ sentry__prepare_event(const sentry_options_t *options, sentry_value_t event,
             envelope, attachment->path, "attachment");
         if (!item) {
             continue;
+        }
+        if (attachment->type != ATTACHMENT) { // don't need to set the default
+            sentry__envelope_item_set_header(item, "attachment_type",
+                sentry_value_new_string(
+                    str_from_attachment_type(attachment->type)));
+        }
+        if (attachment->content_type) {
+            sentry__envelope_item_set_header(item, "content_type",
+                sentry_value_new_string(attachment->content_type));
         }
         sentry__envelope_item_set_header(item, "filename",
 #ifdef SENTRY_PLATFORM_WINDOWS